requirements.txt
glcontext==2.3.7
glfw==2.5.9
moderngl==5.8.2
multipledispatch==0.6.0
numpy==1.24.3
objloader==0.2.0
opencv-python==4.7.0.72
Pillow==9.5.0
pyassimp==4.1.4
pygame==2.4.0
pyglet==2.0.7
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyrr==0.10.3
PyWavefront==1.3.3
six==1.16.0
